refactor: create package for value objects in domain in order to split domain and the vos
