+++ src/main/kotlin/com/barbzdev/portfolio/domain/JobData.kt
@@ -1,5 +1,9 @@
 package com.barbzdev.portfolio.domain
 
+import com.barbzdev.portfolio.domain.valueobject.Link
+import com.barbzdev.portfolio.domain.valueobject.Role
+import com.barbzdev.portfolio.domain.valueobject.Tag
+
 data class JobData(
   val roles: List<Role>,
   val links: List<Link>?,
