--- src/main/kotlin/com/barbzdev/portfolio/domain/JobEndDate.kt
@@ -1,3 +1,0 @@
-package com.barbzdev.portfolio.domain
-
-data class JobEndDate(val value: AuditableDate)
