--- src/main/kotlin/com/barbzdev/portfolio/domain/JobStartDate.kt
@@ -1,3 +1,0 @@
-package com.barbzdev.portfolio.domain
-
-data class JobStartDate(val value: AuditableDate)
