--- src/main/kotlin/com/barbzdev/portfolio/domain/JobUpdatedAt.kt
@@ -1,3 +1,0 @@
-package com.barbzdev.portfolio.domain
-
-data class JobUpdatedAt(val value: AuditableDateTime)
