--- src/main/kotlin/com/barbzdev/portfolio/domain/RoleEndDate.kt
@@ -1,3 +1,0 @@
-package com.barbzdev.portfolio.domain
-
-data class RoleEndDate(val value: AuditableDate)
