--- src/main/kotlin/com/barbzdev/portfolio/domain/RoleStartDate.kt
@@ -1,3 +1,0 @@
-package com.barbzdev.portfolio.domain
-
-data class RoleStartDate(val value: AuditableDate)
