+++ src/main/kotlin/com/barbzdev/portfolio/domain/valueobject/JobUpdatedAt.kt
@@ -1,0 +1,5 @@
+package com.barbzdev.portfolio.domain.valueobject
+
+import com.barbzdev.portfolio.domain.common.AuditableDateTime
+
+data class JobUpdatedAt(val value: AuditableDateTime)
