+++ src/main/kotlin/com/barbzdev/portfolio/domain/valueobject/RoleDescription.kt
@@ -1,4 +1,4 @@
-package com.barbzdev.portfolio.domain
+package com.barbzdev.portfolio.domain.valueobject
 
 data class RoleDescription(val value: String) {
   init {
