+++ src/main/kotlin/com/barbzdev/portfolio/domain/valueobject/RoleEndDate.kt
@@ -1,0 +1,5 @@
+package com.barbzdev.portfolio.domain.valueobject
+
+import com.barbzdev.portfolio.domain.common.AuditableDate
+
+data class RoleEndDate(val value: AuditableDate)
