+++ src/main/kotlin/com/barbzdev/portfolio/domain/valueobject/RoleName.kt
@@ -1,4 +1,4 @@
-package com.barbzdev.portfolio.domain
+package com.barbzdev.portfolio.domain.valueobject
 
 data class RoleName(val value: String) {
   init {
