+++ src/main/kotlin/com/barbzdev/portfolio/domain/valueobject/Tag.kt
@@ -1,4 +1,4 @@
-package com.barbzdev.portfolio.domain
+package com.barbzdev.portfolio.domain.valueobject
 
 
 data class Tag(val value: String) {
